requirements.txt
flask
tensorflow
gdown
flask-cors
numpy==1.24.0
Pillow